// bai_tap_lop/FE/src/features/teacher/landingpage/types.ts
export type FeatureNode = {
  copy: string;
  demo: string;
  image: string;
  imageAlt: string;
  index: number;
  left: string;
  planet: string;
  progress: number;
  title: string;
  top: string;
  theme: "orange" | "blue";
};

export type ModalPosition = {
  connectorX: number;
  left: number;
  top: number;
};